move handler to configuration
